feat: implement AdminUserPage class and remove legacy admin.page.ts
